+++ packages/workbench/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @computesdk/workbench
 
+## 31.0.9
+
+### Patch Changes
+
+- Updated dependencies [deca747]
+  - @computesdk/namespace@1.6.18
+
+## 31.0.8
+
+### Patch Changes
+
+- Updated dependencies [febe062]
+  - @computesdk/blaxel@1.6.22
+
 ## 31.0.7
 
 ### Patch Changes